Move row calculation to timeline worker
